Add 'base' module
