+++ tests/utils.py
@@ -18,6 +18,7 @@ from lz.functional import compose
 from lz.iterating import interleave
 
 from dendroid import (avl,
+                      base,
                       binary,
                       red_black,
                       splay)
@@ -26,9 +27,9 @@ from dendroid.hints import (Domain,
 from dendroid.utils import to_balanced_tree_height
 
 AnyNode = TypeVar('AnyNode', binary.Node, avl.Node, red_black.Node, splay.Node,
-                  binary.NIL)
+                  base.NIL)
 Strategy = SearchStrategy
-Tree = binary.TreeBase
+Tree = base.TreeBase
 TreesPair = Tuple[Tree, Tree]
 TreesTriplet = Tuple[Tree, Tree, Tree]
 ValuesListWithKey = Tuple[List[Domain], Optional[SortingKey]]
@@ -65,17 +66,17 @@ def to_tree_including_value(tree: Tree, value: Domain) -> Tree:
 
 
 def is_left_subtree_less_than_right_subtree(tree: Tree) -> bool:
-    if tree.root is binary.NIL:
+    if tree.root is base.NIL:
         return True
     queue = [(tree.root, tree._to_key(tree.min()), tree._to_key(tree.max()))]
     while queue:
         node, left_end, right_end = queue.pop()
-        if node.left is not binary.NIL:
+        if node.left is not base.NIL:
             if left_end <= node.left.key < right_end:
                 queue.append((node.left, left_end, node.key))
             else:
                 return False
-        if node.right is not binary.NIL:
+        if node.right is not base.NIL:
             if node.key < node.right.key <= right_end:
                 queue.append((node.right, node.key, right_end))
             else:
@@ -93,9 +94,9 @@ def is_node_parent_to_children(node: Union[avl.Node, red_black.Node]) -> bool:
     return _is_child_node(node.left, node) and _is_child_node(node.right, node)
 
 
-def _is_child_node(node: Union[avl.Node, red_black.Node, binary.NIL],
+def _is_child_node(node: Union[avl.Node, red_black.Node, base.NIL],
                    parent: Union[avl.Node, red_black.Node]) -> bool:
-    return node is binary.NIL or node.parent is parent
+    return node is base.NIL or node.parent is parent
 
 
 def to_height(tree: Tree) -> int:
@@ -112,7 +113,7 @@ to_min_binary_tree_height = compose(to_balanced_tree_height, len)
 
 
 @singledispatch
-def to_max_binary_tree_height(tree: binary.TreeBase) -> int:
+def to_max_binary_tree_height(tree: Tree) -> int:
     raise TypeError('Unsupported tree type: {type}.'
                     .format(type=type(tree)))
 
@@ -172,30 +173,30 @@ def to_black_nodes_count(path: Sequence[red_black.Node]) -> int:
 
 
 def iter_nodes(root: AnyNode) -> Iterable[AnyNode]:
-    if root is binary.NIL:
+    if root is base.NIL:
         return
     queue = [root]
     while queue:
         node = queue.pop()
         yield node
-        if node.left is not binary.NIL:
+        if node.left is not base.NIL:
             queue.append(node.left)
-        if node.right is not binary.NIL:
+        if node.right is not base.NIL:
             queue.append(node.right)
 
 
 def to_paths_to_leaves(root: AnyNode) -> Iterable[Sequence[AnyNode]]:
-    if root is binary.NIL:
+    if root is base.NIL:
         return
     queue = [[root]]
     while queue:
         path = queue.pop()
         last_node = path[-1]
         ended = True
-        if last_node.left is not binary.NIL:
+        if last_node.left is not base.NIL:
             ended = False
             queue.append(path + [last_node.left])
-        if last_node.right is not binary.NIL:
+        if last_node.right is not base.NIL:
             ended = False
             queue.append(path + [last_node.right])
         if ended:
